+++ src/app/funding/list/[id]/page.tsx
@@ -1,9 +1,11 @@
 "use client";
-import BottomFixedButton from "@/app/(_components)/bottom-fixed-button";
+import BottomFixedButton, {
+  Button,
+} from "@/app/(_components)/bottom-fixed-button";
 import FundingStatus from "@/app/(_components)/funding-status";
 import { data } from "../../(_dummy)/detail-data";
 import { useState } from "react";
-import InputModal from "./input-modal";
+import BottomSheet from "../../../(_components)/bottom-sheet";
 
 export default function Page() {
   // const {id} = await params;
@@ -63,7 +65,22 @@ export default function Page() {
       <BottomFixedButton onClick={() => setOpen(true)}>
         기부하기
       </BottomFixedButton>
-      <InputModal isOpen={isOpen} setOpen={setOpen} />
+      <BottomSheet isOpen={isOpen} setOpen={setOpen} snapPoints={[300]}>
+        <form className="h-full flex flex-col justify-between">
+          <label htmlFor="funding_amount" className="text-t3">
+            기부 금액
+          </label>
+          <div className="flex gap-2 items-center ">
+            <input
+              id="funding_amount"
+              type="number"
+              className="flex-1 outline-0 text-bd1 bg-transparent border-b border-shadow-600 focus:border-accent-green text-center tracking-wide"
+            />
+            <span>원</span>
+          </div>
+          <Button>확인</Button>
+        </form>
+      </BottomSheet>
     </>
   );
 }
